requirements.txt
keras==2.12.0
numpy==1.23.5
opencv_python==4.7.0.72
Pillow==9.5.0
Pillow==9.5.0
pymongo==4.3.3
Requests==2.29.0
streamlit==1.21.0
streamlit_option_menu==0.3.2
